requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
aiogram==3.4.1
asyncpg==0.29.0
redis==5.0.1
aiosqlite==0.19.0